Implement UserBrickPanel component
