+++ osu.Game.Tests/Visual/Online/TestSceneUserPanel.cs
@@ -42,6 +42,19 @@ namespace osu.Game.Tests.Visual.Online
                 Spacing = new Vector2(10f),
                 Children = new Drawable[]
                 {
+                    new UserBrickPanel(new User
+                    {
+                        Username = @"flyte",
+                        Id = 3103765,
+                        CoverUrl = @"https://osu.ppy.sh/images/headers/profile-covers/c6.jpg"
+                    }),
+                    new UserBrickPanel(new User
+                    {
+                        Username = @"peppy",
+                        Id = 2,
+                        Colour = "99EB47",
+                        CoverUrl = @"https://osu.ppy.sh/images/headers/profile-covers/c3.jpg",
+                    }),
                     flyte = new UserGridPanel(new User
                     {
                         Username = @"flyte",
